Move constants into config
