+++ ui/src/KaptureApp/api/kapture.js
@@ -6,7 +6,7 @@ import {
   COMPONENT_TYPE_COMPOUND,
   COMPONENT_TYPE_MEDIUM,
   COMPONENT_TYPE_SUPPLEMENT,
-} from '../constants';
+} from 'KaptureApp/config/constants';
 import { plateFunctions } from '../store';
 const { createComponent } = plateFunctions;
